use a single ts_def field for aliases, ranges, and floats
For every user of the API who is not concerned with numeric types,
the change should be simple: (Some ty) becomes (Alias ty), and
both (Range _) and (Float _) can be treated as None.

At the late stages of the transformation chain (e.g., during
polymorphism encoding), Range and Float are largely irrelevant:
if these types are supported by the target prover, then the printer
will know what to do with them, and if not, eliminate_literal will
eliminate all Tconst's of this type, and the type can be treated as
abstract.