To amend
Update configure and regtests / nightly-build so that realization are
generated again.

* Makefile.in
POSSIBLE REGRESSION: realization for Isabelle are not generated during the
general make. This allows us to treat it as with coq and be able to detect
changes in realizations with a script.

* configure.in
Adding option --enable-isabelle-no-install in configure which allow people
not wanting to install isabelle to check that they did not change isabelle
realization using make update-isabelle.

* examples/regtests.sh
Adding option --check-realization and --only-realization which add a test
the compares the output of the new realization compared to the one
recorded. It complains if they are different forcing the user to update the
realization files.

* lib/isabelle/*
Adding xml files corresponding to generated realization for Isabelle. This
allows for the abovementionned script to work even if Isabelle is not
installed (which is the case for most people editing theories).

* examples/nightly-bench.sh
Add the option for realization check