[WIP] suggestion for #114. Also: incomplete patch (do not merge)
Relevant file is Ident.ml.
The idea is to add a new type "name" which is an abstract for string.
It is used in the field id_string of ident.
This way, it is not possible to create new unusable names through
appending/prepending keywords like "_spec" or "_def". The function
name_concat_append (etc) provide ways to create new string that can
be used (removing the "infix " part when appending for example).
An unsafe part is still usable: the identity: name -> string is
provided by Ident.
Note that "name" could also be an algebraic type.
type name = | Infix of string | Prefix of string | ...
Note also that this raises some problems with data structures containing
both elements of name type and string type.